checkasm: Fix comparison of different int signs
Casting here should be safe as the case name_length < 0 is already
checked before.

+++ tests/checkasm/checkasm.c
@@ -474,7 +474,7 @@ void *checkasm_check_func(void *const func, const char *const name, ...) {
     const int name_length = vsnprintf(name_buf, sizeof(name_buf), name, arg);
     va_end(arg);
 
-    if (!func || name_length <= 0 || name_length >= sizeof(name_buf))
+    if (!func || name_length <= 0 || (size_t)name_length >= sizeof(name_buf))
         return NULL;
 
     state.current_func = get_func(&state.funcs, name_buf);
